% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_char_mat.R
\name{create.char.mat}
\alias{create.char.mat}
\title{Create Character Matrix}
\usage{
create.char.mat(nrows, ncols)
}
\arguments{
\item{nrows:}{Number of rows}

\item{ncols:}{Number of columns}
}
\value{
Returns character matrix
}
\description{
Creates a matrix whose characters correspond to the labeling of the laboratory system.
}
