% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_positives_vector.R
\name{create_positives_vector}
\alias{create_positives_vector}
\title{Create Vector of positive Items}
\usage{
create_positives_vector(n, mode = "iid", output = "VL", p = 0.01)
}
\arguments{
\item{n:}{Number of items}

\item{mode:}{'iid' for i.i.d. positives or 'fixed_d' for a fixed number of
positives in random positions}

\item{output:}{Either a 'binary' vector or a vector with Viral loads ('VL')}

\item{p:}{Probability of item being positive. For i.i.d. mode p is taken as
probability and for fixed d mode d is taken to be ceil(p * n)}
}
\value{
Returns a vector of 0s and either 1s ('binary') or viral loads ('vL')
}
\description{
Takes in number of items and either a probability of an item being positive
or a total number of positives. Returns a length n 0-1 or 0-VL vector of negatives
and positives.
}
\details{
Functionalities for changing VL generation to be implemented later.
}
