% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder_disjunct_NCOMP.R
\name{decoder_disjunct_NCOMP}
\alias{decoder_disjunct_NCOMP}
\title{Decodes disjunct matrices using NCOMP}
\usage{
decoder_disjunct_NCOMP(n, d, M, Y, NCOMP.tolerance = 0)
}
\arguments{
\item{n:}{Number of subjects}

\item{k:}{disjunctness}

\item{M:}{Test Matrix}

\item{Y:}{Output vector}

\item{NCOMP.tolerance:}{Tolerance of decoder determining number of tests
an item is in that are allowed to be negative while still declaring the item
positive. Increases false positive but decreases false negative rate.}
}
\value{
Returns a vector of inferred positive items (1 := item is positive, 0 := item is negative)
}
\description{
Returns the positive items using NCOMP decoder [1], where positive items are defined by being in at
least d-NCOMP.tolerance tests (where d is the disjunctness of the
test matrix).
}
\references{
Chan, CL et al. Non-adaptive probabilistic group testing with
noisy measurements: Near-optimal bounds with efficient algorithms. In:
Forty-Ninth Annual Allerton Conference Allerton House, UIUC, Illinois, USA
September 28 - 30, 2011.
}
