% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_Y.R
\name{determine_Y}
\alias{determine_Y}
\title{Determine experiment result vector}
\usage{
determine_Y(M, X, mode = "Ct", Se = 1, Sp = 1, spread = 4)
}
\arguments{
\item{M:}{Testing Matrix of 1s and 0s}

\item{X:}{Vector with either positives as 1s or with individual Viral Loads}

\item{mode:}{'binary' returns binary outcome vector (test positive or negative),
'Ct' returns Ct-Valued outcome vector}

\item{Se:}{Test Sensitivity (not implemented for 'Ct)}

\item{Sp:}{Test Specificity (not implemented for 'Ct)}

\item{spread:}{The maximum deviation of a Ct-Value from its expected value.}
}
\value{
Returns test result vector in form specified via input
}
\description{
Determine the experiment result vector (either in binary or Ct-Values) from
testing matrix M and input vector (samples) X. Only works for constant
weight codes at the moment.
}
