% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_STD_testing_protocol.R
\name{generate_STD_testing_protocol}
\alias{generate_STD_testing_protocol}
\title{Shifted Transversal Design - Generate Test Matrix - ADAPTED FOR LAB TESTING}
\usage{
generate_STD_testing_protocol(n, d, control = FALSE, sq = c(7, 0, 1, 4))
}
\arguments{
\item{n:}{number of subjects/items (preferably square number)}

\item{d:}{desired disjunctiveness}

\item{control:}{T or F, turns automatically to TRUE if n isn't a square.}
}
\value{
Returns a d-disjunct Test Matrix from STD design with n columns and (d+1) * sqrt(n) rows (if control == F).
}
\description{
Generates the test matrix for the shifted transversal design [1].
Columns are subjects and rows are tests. Requires the number of subjects
to be a square number. If not, the next lowest square number will be taken
for the design and all the left over subjects will be grouped in one "control"
group at the end.
}
\references{
[1] A new pooling strategy for high-throughput screening:
the Shifted Transversal Design. Thierry-Mieg, N. (2006), BMC Bio Inf, 7:28,
doi: 10.1186/1471-2105-7-28
}
