% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL_plus_prior.R
\name{LL_plus_prior}
\alias{LL_plus_prior}
\title{Optimizes Ct-Values of given status vector}
\usage{
LL_plus_prior(
  gt_model,
  status_vec,
  M,
  Y,
  fittedpars,
  sigma = 0.5,
  out_val = "SE",
  ctbase = 50,
  ctmult = 1
)
}
\arguments{
\item{stan_model:}{Compiled Stan model to be used}

\item{status_vec:}{Binary status vector of positive and negative items}

\item{M:}{Test Matrix}

\item{Y:}{Output vector with Ct-Values}

\item{fittedpars:}{The parameters of the Ct-Value prior distribution}

\item{sigma:}{The standard variation we assume Ct-Values can have. Preset to .5}
}
\value{
Returns the squared average error of the optimization process
}
\description{
Optimizes Ct-Values of given status vector using a stan optimizer with a
given stan model using LL and a prior over Ct-Values.
}
