% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_decoder.R
\name{run_decoder}
\alias{run_decoder}
\title{Optimizes Ct-Values of given status vector}
\usage{
run_decoder(
  Y,
  M,
  n,
  l,
  fittedpars = list(28.50433862, 6.664942792, 28.50433862, 6.664942792),
  sigma_est = 0.5,
  out_val = "SE",
  sum_iterations = 20,
  sum_threshold = 10,
  ctbase = 50,
  ctmult = 1,
  ctmindetect = 25
)
}
\arguments{
\item{Y:}{Group Test Ct-values}

\item{M:}{Design Matrix}

\item{n:}{Number of samples}

\item{l:}{Disjunctness of design matrix}

\item{fittedpars:}{The parameters of the Ct-Value prior distribution. Provide as list of (mean, sd, mean, sd). Preset to list(28.50433862,6.664942792,28.50433862,6.664942792)}

\item{sigma_est:}{The standard variation we assume Ct-Values can have. Preset to .5}

\item{out_val:}{Error used in simulated annealing. Default 'SE' (squared error)}

\item{sum_iterations:}{Number of times decoder is run. Default: 20}

\item{sum_threshold:}{Decision Threshold as intege. Default: 10}

\item{ctbase:}{Ct-Value generation Baseline. Default: 50}

\item{ctmult:}{Multiplier for VLs in Ct-Value generation. Default: 1}

\item{ctmindetect:}{Ct-value for potentially hidden samples. Default: 25}
}
\value{
Returns the indeces of the positive sampless
}
\description{
Optimizes Ct-Values of given status vector using a stan optimizer with a
given stan model using LL and a prior over Ct-Values.
}
