% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_test.R
\name{run_test}
\alias{run_test}
\title{Runs a test for the decoder after installation}
\usage{
run_test()
}
\description{
Creates a test design, vectors of positives and group test results and runs the decoder.
}
