% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VL_generator.R
\name{VL_generator}
\alias{VL_generator}
\title{Generate fake Viral Loads}
\usage{
VL_generator(
  num_samples,
  distribution,
  alpha = 7.98,
  t_w = 15.1,
  t_inc = 5.52,
  t_g = 0,
  t_p = 3,
  t_inf = 0,
  lower_bound = 0,
  upper_bound = 1e+08
)
}
\arguments{
\item{num_samples:}{Number of positive swap samples to be taken}

\item{distribution:}{'Uniform' or 'Cleary'}

\item{max_time:}{last time a sample can be taken}

\item{alpha...:}{See Cleary et al., 2020}

\item{lower_bound:}{Lower bound for uniform distribution}

\item{upper_bound:}{Upper bound for uniform distribution}
}
\value{
Returns vector of viral loads of length n.
}
\description{
Generates fake viral loads randomly either from a uniform distribution or
by the process described by Cleary [1].
}
\references{
[1] Brian Cleary et al. \(2020\), Using viral load and epidemic dynamics to optimize
pooled testing in resource constrained settings. medRxiv, doi: 10.1101/2020.05.01.20086801,
version from October 6, 2020
}
