% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VL_to_Ct.R
\name{VL_to_Ct}
\alias{VL_to_Ct}
\title{Viral Load to Ct-Value}
\usage{
VL_to_Ct(
  VL,
  mode = "range",
  spread = 2,
  baseline_Ct = 37,
  baseline_VL = 1,
  det = FALSE
)
}
\arguments{
\item{VL:}{viral Load number.}

\item{mode:}{'range' (for choosing noise from uniform range) or 'dist' for
choosing noise via a distribution.}

\item{spread:}{spread of Ct-values, either a range for mode == 'range'
or a standard deviation for mode == 'dist'.}

\item{baseline_Ct:}{baseline Ct at baseline Viral Load as starting point
for dilution based approximation}

\item{baseline_VL:}{Baseline Viral load generating baseline Ct as starting point
for dilution based approximation}
}
\value{
Returns a Ct-Value
}
\description{
Converts a Viral Load into a Ct-value, infusing noise either from
a uniform spread or a distribution. Default values are loosely taken from
data in [1].
}
\note{
Doesn't allow for false positives.
}
\references{
[1] Chantal BF Vogels et al. Analytical sensitivity and efficiency
comparisons of SARS CoV 2 RT qPCR primer probe sets. In:
Naturemicrobiology 5.10 (2020), pp. 1299 to 1305.
}
